basic boost stats work

+++ TrackedStats/TrackedStat.h
@@ -3,6 +3,6 @@
 struct TrackedStat
 {
 	std::string name_;
-	std::function<int()> getter_;
+	std::function<int64_t()> getter_;
 };
